apache-cassandra-data-modeling/requirements.txt
psycopg2-binary
pandas
numpy
jupyter
ipython-sql
cassandra-driver